Move volume/store to component/volume/store
